Fix for Mac: Ajustment for Mac bash

+++ scripts/shared/functions.sh
@@ -3,8 +3,12 @@
 function isJson() {
   local maybeJson=$1
   
-  echo "$maybeJson" | jq -e >/dev/null 2>&1
-  
+  if [[ -f "$maybeJson" ]]; then
+    jq -e '.' "$maybeJson" >/dev/null 2>&1
+  else
+    echo "$maybeJson" | jq -e >/dev/null 2>&1
+  fi
+
   if [[ $? -gt 0 ]]; then
     echo 1
     return 1
@@ -12,4 +16,14 @@ function isJson() {
     echo 0
     return 0
   fi
+}
+
+function randomAlphaNumeric() {
+  if [[ $# -gt 0 &&  $(echo "$1" | grep -Ec '^[0-9]+$') == 1 ]]; then
+    local length=$1
+  else 
+    local length=40
+  fi
+
+  tr -dc A-Za-z0-9 </dev/urandom | head -c $length; echo
 }
